Updates on PARSM

+++ GDPARSM_Main_PASCUAL/FPSCounter.h
@@ -14,7 +14,7 @@ class FPSCounter :    public AGameObject
 	private:
 		sf::Time updateTime;
 		sf::Text* statsText;
-		int framesPassed = 0;
+		float framesPassed = 0;
 
 		void updateFPS(sf::Time elapsedTime);
